% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastPCA_base}
\alias{fastPCA_base}
\title{Prepare the fastPCA result object}
\usage{
fastPCA_base(result, k)
}
\arguments{
\item{result}{Value of one of the other functions.}

\item{k}{Rank of the transformation.}
}
\value{
A FastPCA object, containing the decomposition.
}
\description{
This is an internal function that formats the resulting fastPCA object so that it can be handled appropriately by R
}

