% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastPCA_BED}
\alias{fastPCA_BED}
\title{Perform fast SVD for a matrix in BED format}
\usage{
fastPCA_BED(inputFile, k = 5, l, mem, its = 2, phenoFile = -1,
  diffsnorm = FALSE, centeringRow = FALSE, centeringColumn = FALSE)
}
\arguments{
\item{inputFile}{Base file name for the .bed,.fam, .bim files}

\item{k}{Rank of decomposition. Default: 5}

\item{l}{Block size. Default k+2}

\item{its}{Number of normalized power iterations. Default: 2}

\item{diffsnorm}{Calculate 2-norm accuracy, i.e. ||A-USV||_2.}

\item{centeringRow}{Center the rows prior to decomposition.}

\item{centeringRow}{Center the columns prior to decomposition.}
}
\value{
A FastPCA object, containing the decomposition.
}
\description{
This function performs a nearly optimal rank-k approximation to the singular value decomposition inputMatrix = USV' on a matrix that is passed via BED format.  Please see references for explanation of 'nearly optimal.'
}

