\name{fastRPCA-package}
\alias{fastRPCA-package}
\alias{fastRPCA}
\docType{package}
\title{
\packageTitle{fastRPCA}
}
\description{
\packageDescription{fastRPCA}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fastRPCA}
\packageIndices{fastRPCA}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{fastRPCA}

Maintainer: \packageMaintainer{fastRPCA}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
